Fix compilation problem in SipProxy::SipProxy, uint16_t vs uint16_t& mismatch.
Return nread in RtpPacket::internalReadPacket, which didn't return any value,
and fix start of data in RtpPacket::readPacketFrom and
SRtpPacket::readPacketFrom.

/Mikael


git-svn-id: svn://svn.minisip.org/minisip/trunk@1712 63dae60f-c5d3-0310-92a3-fcbf00261c72

+++ libmsip/source/SipDialogConfig.cxx
@@ -58,13 +58,13 @@ SipProxy::SipProxy(std::string addr, int port){
 
 SipProxy::SipProxy(std::string userUri, string transport) {
 	string addr;
-	int port;
+	uint16_t port;
 	autodetectSettings = true;
 	
 	registerExpires=DEFAULT_SIPPROXY_EXPIRES_VALUE_SECONDS;
 	defaultExpires=DEFAULT_SIPPROXY_EXPIRES_VALUE_SECONDS;
 	try {
-		addr = SipProxy::findProxy( userUri, (uint16_t)port, transport );
+		addr = SipProxy::findProxy( userUri, port, transport );
 	}catch( NetworkException * exc ) {
 		if( transport == "TCP" ) { 
 			//if tcp doesn't work, try find UDP
@@ -77,7 +77,7 @@ SipProxy::SipProxy(std::string userUri, string transport) {
 		try {
 			cerr << "Autodetect Sip proxy for [" << userUri << "] for transport TCP failed. Retrying with transport UDP." << endl;
 			transport = "UDP";
-			addr = SipProxy::findProxy( userUri, (uint16_t)port, transport );
+			addr = SipProxy::findProxy( userUri, port, transport );
 		}catch( NetworkException * exc2 ) {
 				addr = "unknown";
 		}
